feat(dashboard): shared MergeButton and RecoverButton, organize inspector by workspace/issue
- Extract MergeButton and RecoverButton as shared components with
  variant='card' | 'inspector' styling
- Reorganize inspector ActionsSection into Workspace and Issue sections
  with visual headers (Box/Tag icons)
- Add LayoutGrid icon indicator next to inspector actions that also
  appear on kanban cards
- Normalize all action labels between card and inspector variants
- Remove inline MergeIssueButton and ResetPipelineButton from KanbanBoard
