+++ src/dashboard/frontend/src/components/inspector/ActionsSection.tsx
@@ -1,6 +1,6 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import {
-  XCircle, RefreshCw, CheckCircle, Play, FolderPlus, Check, Loader2, RotateCcw, X, Send, AlertTriangle, ChevronRight,
+  XCircle, RefreshCw, CheckCircle, Play, FolderPlus, Check, Loader2, RotateCcw, X, Send, ChevronRight, Box, Tag, LayoutGrid,
 } from 'lucide-react';
 import type { UseMutationResult } from '@tanstack/react-query';
 import { Agent, WorkAgentLifecycle } from '../../types';
@@ -9,10 +9,12 @@ import { ReviewPipelineSection } from './ReviewPipelineSection';
 import { isReviewPipelineStuck } from '../../lib/pipeline-state';
 import { ResetIssueButton } from '../ResetIssueButton';
 import { StopAgentButton } from '../StopAgentButton';
+import { MergeButton } from '../MergeButton';
+import { RecoverButton } from '../RecoverButton';
+import { ArtifactLinks } from '../ArtifactLinks';
 
 // Convenience alias — most mutations use void variables and unknown data
 type AnyMutation = UseMutationResult<unknown, Error, void, unknown>;
-type ResetReviewMutation = UseMutationResult<unknown, Error, { rerun?: boolean } | undefined, unknown>;
 type SyncMutation = UseMutationResult<{ alreadyUpToDate?: boolean; commitCount?: number }, Error, void, unknown>;
 type ReopenMutation = UseMutationResult<unknown, Error, string | undefined, unknown>;
 
@@ -22,25 +24,25 @@ interface ActionsSectionProps {
   reviewStatus?: ReviewStatus;
   reviewStatusLoading?: boolean;
   workspace?: WorkspaceInfo;
-  mergeMutation: AnyMutation;
+  hasPlan: boolean;
+  beadsCount: number;
   reviewMutation: AnyMutation;
   cancelMutation: AnyMutation;
-  resetReviewMutation: ResetReviewMutation;
   startAgentMutation: UseMutationResult<unknown, Error, string | undefined, unknown>;
   createWorkspaceMutation: AnyMutation;
   syncMainMutation: SyncMutation;
   resetSessionMutation: AnyMutation;
   reopenMutation?: ReopenMutation;
-  onMerge: () => void;
   onReview: () => void;
   onCancel: () => void;
-  onResetReview: () => void;
   onResetSession: () => void;
   onDismissPending: () => void;
   onStartAgent: (message?: string) => void;
   onCreateWorkspace: () => void;
   onReopen?: () => void;
   onKillSuccess?: () => void;
+  onViewBeads: () => void;
+  onViewVBrief: () => void;
   lifecycle?: WorkAgentLifecycle;
   agentLaunchState?: 'starting' | 'resuming' | null;
 }
@@ -51,25 +53,25 @@ export function ActionsSection({
   reviewStatus,
   reviewStatusLoading,
   workspace,
-  mergeMutation,
+  hasPlan,
+  beadsCount,
   reviewMutation,
   cancelMutation,
-  resetReviewMutation,
   startAgentMutation,
   createWorkspaceMutation,
   syncMainMutation,
   resetSessionMutation,
   reopenMutation,
-  onMerge,
   onReview,
   onCancel,
-  onResetReview,
   onResetSession,
   onDismissPending,
   onStartAgent,
   onCreateWorkspace,
   onReopen,
   onKillSuccess,
+  onViewBeads,
+  onViewVBrief,
   lifecycle,
   agentLaunchState,
 }: ActionsSectionProps) {
@@ -79,18 +81,6 @@ export function ActionsSection({
   const isLaunching = agentLaunchState === 'starting' || agentLaunchState === 'resuming';
   const launchLabel = agentLaunchState === 'resuming' ? 'Resuming...' : 'Starting...';
 
-  // Stuck merge detection: if mergeStatus has been 'merging' for > 2 min, enable retry (PAN-490)
-  const STUCK_MERGE_MS = 2 * 60 * 1000;
-  const [now, setNow] = useState(() => Date.now());
-  useEffect(() => {
-    if (reviewStatus?.mergeStatus !== 'merging') return;
-    const interval = setInterval(() => setNow(Date.now()), 30_000);
-    return () => clearInterval(interval);
-  }, [reviewStatus?.mergeStatus]);
-  const mergingElapsed = reviewStatus?.mergeStatus === 'merging' && reviewStatus.updatedAt
-    ? now - new Date(reviewStatus.updatedAt).getTime()
-    : 0;
-  const isMergeStuck = mergingElapsed > STUCK_MERGE_MS;
   const isPipelineStuck = isReviewPipelineStuck(reviewStatus);
   const hasVerificationState = !!reviewStatus?.verificationStatus && reviewStatus.verificationStatus !== 'pending';
   const showPipelineStatus = !!reviewStatus && (
@@ -180,116 +170,108 @@ export function ActionsSection({
         </div>
       )}
 
-      <div className="flex flex-wrap gap-1.5 mt-2">
-        {/* MERGE button */}
-        {reviewStatus?.readyForMerge && reviewStatus?.mergeStatus !== 'merged' && (
+      {/* Workspace Actions */}
+      <div className="mt-4">
+        <div className="text-xs uppercase tracking-wider mb-2 font-semibold text-content-subtle flex items-center gap-1.5">
+          <Box className="w-3 h-3" /> Workspace
+        </div>
+        <div className="flex flex-wrap gap-1.5">
+          {/* MERGE */}
+          <div className="flex items-center gap-1">
+            <MergeButton issueId={issueId} reviewStatus={reviewStatus} variant="inspector" />
+            <span title="Also shown on card">
+              <LayoutGrid className="w-3 h-3 text-content-subtle opacity-40 self-center" />
+            </span>
+          </div>
+          {reviewStatus?.mergeStatus === 'merged' && (
+            <span className="flex items-center gap-1 px-2 py-1 text-xs badge-bg-success text-success rounded font-medium">
+              <CheckCircle className="w-3 h-3" />Merged
+            </span>
+          )}
+
+          {/* Review & Test */}
           <button
-            data-testid="merge-btn"
-            onClick={onMerge}
-            disabled={mergeMutation.isPending || ((reviewStatus?.mergeStatus === 'merging' || reviewStatus?.mergeStatus === 'verifying' || reviewStatus?.mergeStatus === 'queued') && !isMergeStuck)}
-            className={`flex items-center gap-1 px-2 py-1 text-xs rounded font-medium ${
-              isMergeStuck
-                ? 'bg-warning text-white hover:bg-warning/90'
-                : 'bg-success text-white hover:bg-success/90 disabled:opacity-50'
+            data-testid="review-test-btn"
+            onClick={onReview}
+            disabled={reviewMutation.isPending || reviewStatus?.reviewStatus === 'reviewing' || reviewStatus?.testStatus === 'testing'}
+            className={`flex items-center gap-1 px-2 py-1 text-xs rounded disabled:opacity-50 ${
+              shouldPromoteReviewAction
+                ? 'bg-primary text-primary-foreground hover:bg-primary/90 font-medium shadow-sm'
+                : 'text-muted-foreground hover:text-foreground hover:bg-accent'
             }`}
-            title={isMergeStuck ? 'Merge appears stuck — click to retry' : undefined}
           >
-            {mergeMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> :
-             isMergeStuck ? <AlertTriangle className="w-3 h-3" /> :
-             reviewStatus?.mergeStatus === 'verifying' ? <Loader2 className="w-3 h-3 animate-spin" /> :
-             reviewStatus?.mergeStatus === 'merging' ? <Loader2 className="w-3 h-3 animate-spin" /> :
-             <CheckCircle className="w-3 h-3" />}
-            {isMergeStuck ? 'RETRY MERGE' :
-             reviewStatus?.mergeStatus === 'queued' ? 'QUEUED' :
-             reviewStatus?.mergeStatus === 'verifying' ? 'VERIFYING...' :
-             reviewStatus?.mergeStatus === 'merging' ? 'REBASING...' : 'MERGE'}
+            {(reviewMutation.isPending || reviewStatus?.reviewStatus === 'reviewing' || reviewStatus?.testStatus === 'testing') ?
+              <Loader2 className="w-3 h-3 animate-spin" /> : <RefreshCw className="w-3 h-3" />}
+            {isReReview ? 'Re-Review' : 'Review & Test'}
           </button>
-        )}
-        {reviewStatus?.mergeStatus === 'merged' && (
-          <span className="flex items-center gap-1 px-2 py-1 text-xs badge-bg-success text-success rounded font-medium">
-            <CheckCircle className="w-3 h-3" />MERGED
-          </span>
-        )}
 
-        {/* Review & Test */}
-        <button
-          data-testid="review-test-btn"
-          onClick={onReview}
-          disabled={reviewMutation.isPending || reviewStatus?.reviewStatus === 'reviewing' || reviewStatus?.testStatus === 'testing'}
-          className={`flex items-center gap-1 px-2 py-1 text-xs rounded disabled:opacity-50 ${
-            shouldPromoteReviewAction
-              ? 'bg-primary text-primary-foreground hover:bg-primary/90 font-medium shadow-sm'
-              : 'text-muted-foreground hover:text-foreground hover:bg-accent'
-          }`}
-        >
-          {(reviewMutation.isPending || reviewStatus?.reviewStatus === 'reviewing' || reviewStatus?.testStatus === 'testing') ?
-            <Loader2 className="w-3 h-3 animate-spin" /> : <RefreshCw className="w-3 h-3" />}
-          {isReReview ? 'Re-Review' : 'Review & Test'}
-        </button>
+          {/* Stop Agent */}
+          {agent && agent.status !== 'stopped' && (
+            <div className="flex items-center gap-1">
+              <StopAgentButton
+                agentId={agent?.id}
+                variant="inspector"
+                onSuccess={onKillSuccess}
+              />
+              <span title="Also shown on card">
+                <LayoutGrid className="w-3 h-3 text-content-subtle opacity-40 self-center" />
+              </span>
+            </div>
+          )}
 
-        {/* Stop Agent */}
-        {agent && agent.status !== 'stopped' && (
-          <StopAgentButton
-            agentId={agent?.id}
-            variant="inspector"
-            onSuccess={onKillSuccess}
-          />
-        )}
+          {/* Recover failed review/test/merge pipeline */}
+          {reviewStatus && isPipelineStuck && (
+            <div className="flex items-center gap-1">
+              <RecoverButton issueId={issueId} reviewStatus={reviewStatus} variant="inspector" />
+              <span title="Also shown on card">
+                <LayoutGrid className="w-3 h-3 text-content-subtle opacity-40 self-center" />
+              </span>
+            </div>
+          )}
 
-        {/* Recover failed review/test/merge pipeline */}
-        {reviewStatus && isPipelineStuck && (
-          <button
-            onClick={onResetReview}
-            disabled={resetReviewMutation.isPending}
-            className="flex items-center gap-1 px-2 py-1 text-xs text-muted-foreground rounded hover:text-foreground hover:bg-accent disabled:opacity-50"
-          >
-            {resetReviewMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : <RotateCcw className="w-3 h-3" />}
-            {resetReviewMutation.isPending ? 'Recovering...' : 'Recover'}
-          </button>
-        )}
-
-        {/* Start/Resume Agent when no agent or stopped */}
-        {(!agent || agent.status === 'stopped') && (
-          <>
-            <button
-              onClick={() => {
-                if (isResume) {
-                  setShowResumeInput(true);
-                } else {
-                  onStartAgent();
-                }
-              }}
-              disabled={isLaunching || showResumeInput}
-              className="flex items-center gap-1 text-xs text-primary hover:text-primary/80 transition-colors disabled:opacity-50"
-            >
-              {isLaunching ? <Loader2 className="w-3 h-3 animate-spin" /> : <Play className="w-3 h-3" />}
-              <span>{isLaunching ? launchLabel : (isResume ? 'Resume Session' : 'Start Agent')}</span>
-            </button>
-            {/* Reset Session — only when resuming (has a saved session) */}
-            {isResume && (
+          {/* Start/Resume Agent when no agent or stopped */}
+          {(!agent || agent.status === 'stopped') && (
+            <>
               <button
-                onClick={onResetSession}
-                disabled={resetSessionMutation.isPending || resetSessionMutation.isSuccess}
-                className="flex items-center gap-1 px-2 py-1 text-xs text-muted-foreground rounded hover:text-foreground hover:bg-accent disabled:opacity-50"
-                title="Clear saved session so next start creates a fresh Claude session (preserves workspace)"
+                onClick={() => {
+                  if (isResume) {
+                    setShowResumeInput(true);
+                  } else {
+                    onStartAgent();
+                  }
+                }}
+                disabled={isLaunching || showResumeInput}
+                className="flex items-center gap-1 text-xs text-primary hover:text-primary/80 transition-colors disabled:opacity-50"
               >
-                {resetSessionMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : resetSessionMutation.isSuccess ? <Check className="w-3 h-3" /> : <RotateCcw className="w-3 h-3" />}
-                {resetSessionMutation.isPending ? 'Resetting...' : resetSessionMutation.isSuccess ? 'Session Reset' : 'Reset Session'}
+                {isLaunching ? <Loader2 className="w-3 h-3 animate-spin" /> : <Play className="w-3 h-3" />}
+                <span>{isLaunching ? launchLabel : (isResume ? 'Resume Session' : 'Start Agent')}</span>
               </button>
-            )}
-            {!workspace?.exists && (
-               <button
-                 onClick={onCreateWorkspace}
-                 disabled={createWorkspaceMutation.isPending || createWorkspaceMutation.isSuccess}
-                 className="flex items-center gap-1 px-2 py-1 text-xs text-white rounded disabled:opacity-50 border bg-surface-emphasis border-divider"
-               >
-                 {(createWorkspaceMutation.isPending || createWorkspaceMutation.isSuccess) ? <Loader2 className="w-3 h-3 animate-spin" /> : <FolderPlus className="w-3 h-3" />}
-                 {createWorkspaceMutation.isPending ? 'Creating...' : 'Create Workspace'}
-               </button>
-             )}
-           </>
-         )}
-       </div>
+              {/* Reset Session — only when resuming (has a saved session) */}
+              {isResume && (
+                <button
+                  onClick={onResetSession}
+                  disabled={resetSessionMutation.isPending || resetSessionMutation.isSuccess}
+                  className="flex items-center gap-1 px-2 py-1 text-xs text-muted-foreground rounded hover:text-foreground hover:bg-accent disabled:opacity-50"
+                  title="Clear saved session so next start creates a fresh Claude session (preserves workspace)"
+                >
+                  {resetSessionMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : resetSessionMutation.isSuccess ? <Check className="w-3 h-3" /> : <RotateCcw className="w-3 h-3" />}
+                  {resetSessionMutation.isPending ? 'Resetting...' : resetSessionMutation.isSuccess ? 'Session Reset' : 'Reset Session'}
+                </button>
+              )}
+              {!workspace?.exists && (
+                 <button
+                   onClick={onCreateWorkspace}
+                   disabled={createWorkspaceMutation.isPending || createWorkspaceMutation.isSuccess}
+                   className="flex items-center gap-1 px-2 py-1 text-xs text-white rounded disabled:opacity-50 border bg-surface-emphasis border-divider"
+                 >
+                   {(createWorkspaceMutation.isPending || createWorkspaceMutation.isSuccess) ? <Loader2 className="w-3 h-3 animate-spin" /> : <FolderPlus className="w-3 h-3" />}
+                   {createWorkspaceMutation.isPending ? 'Creating...' : 'Create Workspace'}
+                 </button>
+               )}
+             </>
+           )}
+         </div>
+      </div>
 
       {!!agent && agent.status === 'stopped' && lifecycle?.reason && (
         <div className="text-xs text-content-subtle mt-2 px-2 py-1 rounded bg-surface-emphasis/40 border border-divider">
@@ -350,11 +332,6 @@ export function ActionsSection({
           {reviewMutation.error instanceof Error ? reviewMutation.error.message : 'Failed to start review'}
         </div>
       )}
-      {mergeMutation.isError && (
-        <div className="text-xs text-destructive badge-bg-destructive px-2 py-1 rounded mt-2">
-          {mergeMutation.error instanceof Error ? mergeMutation.error.message : 'Failed to merge'}
-        </div>
-      )}
       {startAgentMutation.isError && (
         <div className="text-xs text-destructive badge-bg-destructive px-2 py-1 rounded mt-2">
           {startAgentMutation.error instanceof Error ? startAgentMutation.error.message : 'Failed to start agent'}
@@ -371,55 +348,77 @@ export function ActionsSection({
         </div>
       )}
 
-      {/* Danger Zone — destructive actions (collapsed by default) */}
-      {reviewStatus?.mergeStatus !== 'merged' && (
-        <details className="mt-6 rounded border border-destructive/30 group">
-          <summary className="px-3 py-2 bg-destructive/5 rounded cursor-pointer list-none select-none flex items-center gap-1.5 group-open:rounded-b-none group-open:border-b group-open:border-destructive/30">
-            <ChevronRight className="w-3 h-3 text-destructive transition-transform group-open:rotate-90" />
-            <span className="text-xs font-semibold uppercase tracking-wider text-destructive">Danger Zone</span>
-          </summary>
-          <div className="px-3 py-3 space-y-4">
-            {/* Reopen */}
-            {onReopen && (
+      {/* Issue Actions */}
+      <div className="mt-4">
+        <div className="text-xs uppercase tracking-wider mb-2 font-semibold text-content-subtle flex items-center gap-1.5">
+          <Tag className="w-3 h-3" /> Issue
+        </div>
+        <div className="flex flex-wrap gap-1.5">
+          <div className="flex items-center gap-1">
+            <ArtifactLinks
+              issueId={issueId}
+              hasPlan={hasPlan}
+              beadsCount={beadsCount}
+              onViewBeads={onViewBeads}
+              onViewVBrief={onViewVBrief}
+              variant="inspector"
+            />
+            <span title="Also shown on card">
+              <LayoutGrid className="w-3 h-3 text-content-subtle opacity-40 self-center" />
+            </span>
+          </div>
+        </div>
+
+        {/* Danger Zone — destructive actions (collapsed by default) */}
+        {reviewStatus?.mergeStatus !== 'merged' && (
+          <details className="mt-4 rounded border border-destructive/30 group">
+            <summary className="px-3 py-2 bg-destructive/5 rounded cursor-pointer list-none select-none flex items-center gap-1.5 group-open:rounded-b-none group-open:border-b group-open:border-destructive/30">
+              <ChevronRight className="w-3 h-3 text-destructive transition-transform group-open:rotate-90" />
+              <span className="text-xs font-semibold uppercase tracking-wider text-destructive">Danger Zone</span>
+            </summary>
+            <div className="px-3 py-3 space-y-4">
+              {/* Reopen */}
+              {onReopen && (
+                <div className="min-w-0">
+                  <div className="text-xs font-medium text-content">Reopen for more work</div>
+                  <div className="text-[11px] text-content-subtle mt-0.5" title="Moves the issue back to In Progress so the work agent can continue. Keeps the workspace, branch, PR, STATE.md, and all planning artifacts intact.">
+                    Moves the issue back to In Progress. The workspace, branch, PR, STATE.md, and all planning artifacts are preserved.
+                  </div>
+                  <button
+                    onClick={onReopen}
+                    disabled={reopenMutation?.isPending}
+                    className="mt-2 flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium rounded border border-warning/40 text-warning hover:bg-warning hover:text-white transition-colors disabled:opacity-50"
+                    title="Reopen: moves issue to In Progress, keeps workspace + branch + PR + STATE.md + beads"
+                  >
+                    {reopenMutation?.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : <RefreshCw className="w-3 h-3" />}
+                    {reopenMutation?.isPending ? 'Reopening...' : 'Reopen'}
+                  </button>
+                </div>
+              )}
+
+              {/* Reset Issue */}
+              <ResetIssueButton issueId={issueId} variant="danger-zone" />
+
+              {/* Cancel Issue */}
               <div className="min-w-0">
-                <div className="text-xs font-medium text-content">Reopen for more work</div>
-                <div className="text-[11px] text-content-subtle mt-0.5" title="Moves the issue back to In Progress so the work agent can continue. Keeps the workspace, branch, PR, STATE.md, and all planning artifacts intact.">
-                  Moves the issue back to In Progress. The workspace, branch, PR, STATE.md, and all planning artifacts are preserved.
+                <div className="text-xs font-medium text-content">Cancel this issue</div>
+                <div className="text-[11px] text-content-subtle mt-0.5" title="Permanently stops the agent, deletes the workspace and branch (including STATE.md), closes the PR, removes beads, and moves the issue to Canceled. This cannot be undone.">
+                  Permanently stops the agent, deletes the workspace and branch (including STATE.md), closes the PR, and moves the issue to Canceled. This cannot be undone.
                 </div>
                 <button
-                  onClick={onReopen}
-                  disabled={reopenMutation?.isPending}
-                  className="mt-2 flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium rounded border border-warning/40 text-warning hover:bg-warning hover:text-white transition-colors disabled:opacity-50"
-                  title="Reopen: moves issue to In Progress, keeps workspace + branch + PR + STATE.md + beads"
+                  onClick={onCancel}
+                  disabled={cancelMutation.isPending}
+                  className="mt-2 flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium rounded border border-destructive/40 text-destructive hover:bg-destructive hover:text-white transition-colors disabled:opacity-50"
+                  title="Cancel Issue: permanent — stops agent, deletes workspace + branch + STATE.md, closes PR, moves to Canceled"
                 >
-                  {reopenMutation?.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : <RefreshCw className="w-3 h-3" />}
-                  {reopenMutation?.isPending ? 'Reopening...' : 'Reopen'}
+                  {cancelMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : <XCircle className="w-3 h-3" />}
+                  {cancelMutation.isPending ? 'Canceling...' : 'Cancel Issue'}
                 </button>
               </div>
-            )}
-
-            {/* Reset Issue */}
-            <ResetIssueButton issueId={issueId} variant="danger-zone" />
-
-            {/* Cancel Issue */}
-            <div className="min-w-0">
-              <div className="text-xs font-medium text-content">Cancel this issue</div>
-              <div className="text-[11px] text-content-subtle mt-0.5" title="Permanently stops the agent, deletes the workspace and branch (including STATE.md), closes the PR, removes beads, and moves the issue to Canceled. This cannot be undone.">
-                Permanently stops the agent, deletes the workspace and branch (including STATE.md), closes the PR, and moves the issue to Canceled. This cannot be undone.
-              </div>
-              <button
-                onClick={onCancel}
-                disabled={cancelMutation.isPending}
-                className="mt-2 flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium rounded border border-destructive/40 text-destructive hover:bg-destructive hover:text-white transition-colors disabled:opacity-50"
-                title="Cancel Issue: permanent — stops agent, deletes workspace + branch + STATE.md, closes PR, moves to Canceled"
-              >
-                {cancelMutation.isPending ? <Loader2 className="w-3 h-3 animate-spin" /> : <XCircle className="w-3 h-3" />}
-                {cancelMutation.isPending ? 'Canceling...' : 'Cancel Issue'}
-              </button>
             </div>
-          </div>
-        </details>
-      )}
+          </details>
+        )}
+      </div>
     </div>
   );
 }
